+++ src/app/models/category.model.ts
@@ -5,3 +5,13 @@ export class Category {
   categoryDescription?: string;
   categoryVisible: boolean;
 }
+
+// These labels should be used anywhere in the sheets' column-header, prefixed with a `#`
+// For example: "Category Icon (can be any URL) #ICON"
+export enum CategoryCol {
+  id = 'ID',
+  name = 'NAME',
+  icon = 'ICON',
+  description = 'DESCRIPTION',
+  visible = 'VISIBLE',
+}