+++ CHANGELOG.md
@@ -10,4 +10,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## 2022-10-24
+
+### Changed
+
+- Connection between Google Sheets based on column-order replaced by `#tag`s in column-headers.
+  Any sheets without these `#tag`s will not work anymore.
+
+  See the specific required/available `#tag`s in:
+
+  - [`models/category.model.ts`](./src/app/models/category.model.ts)
+  - [`models/sub-category.model.ts`](./src/app/models/sub-category.model.ts)
+  - [`models/offer.model.ts`](./src/app/models/offer.model.ts)
+  - [`models/qa-set.model.ts`](./src/app/models/qa-set.model.ts)
+
+  See [#204](https://github.com/rodekruis/helpful-information/issues/204)
+
+---
+
 ## Start - 2021-03-10